# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "KYS PORTAL",
    "description": "KYS Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "103.224.246.103:3004",
  "basePath": "/",
...
  "paths": {
...
    "/book": {
      "post": {
        "tags": [
          "Books"
        ],
        "description": "Create new book in system",
        "parameters": [
          {
            "name": "attachment",
            "in": "formData",
            "description": "attachment",
            "type": "file"
          },
          {
            "name": "bookName",
            "in": "formData",
            "description": "bookName",
            "type": "string"
          },
          {
            "name": "pricing",
            "in": "formData",
            "description": "pricing",
            "type": "number"
          },
          {
            "name": "bookDetails",
            "in": "formData",
            "description": "bookDetails",
            "type": "string"
          },
          {
            "name": "category",
            "in": "formData",
            "description": "category",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New book is created",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          }
        }
      }
    },
    "/book/deleteById": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Remove Book in system",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New document is created",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          }
        }
      }
    },
    "/book/update": {
      "post": {
        "tags": [
          "Books"
        ],
        "description": "Create new book in system",
        "parameters": [
          {
            "name": "_id",
            "in": "formData",
            "description": "id",
            "type": "string"
          },
          {
            "name": "attachment",
            "in": "formData",
            "description": "attachment",
            "type": "file"
          },
          {
            "name": "bookName",
            "in": "formData",
            "description": "bookName",
            "type": "string"
          },
          {
            "name": "pricing",
            "in": "formData",
            "description": "pricing",
            "type": "number"
          },
          {
            "name": "bookDetails",
            "in": "formData",
            "description": "bookDetails",
            "type": "string"
          },
          {
            "name": "category",
            "in": "formData",
            "description": "category",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New book is created",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          }
        }
      }
    },
...
    "/book/category": {
      "post": {
        "tags": [
          "Books"
        ],
        "description": "Create new category in system",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "create new category in system",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New book is created",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/book/category/deleteById": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Remove Book in system",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New document is created",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          }
        }
      }
    },
    "/book/category/update": {
      "post": {
        "tags": [
          "Books"
        ],
        "description": "Create new book in system",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "create new category in system",
            "schema": {
              "$ref": "#/definitions/UpdateCategory"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New book is created",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Books": {
      "required": [
        "bookName"
      ],
      "properties": {
        "bookName": {
          "type": "string"
        },
        "pricing": {
          "type": "number"
        },
        "bookDetails": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "attachment": {
          "type": "string"
        },
        "createDate": {
          "type": "number"
        },
        "updateDate": {
          "type": "number"
        }
      }
    },
...
    "UpdateCategory" : {
      "required": [
        "_id"
      ],
      "properties": {
      "value": {
        "type": "string"
      }, 
      "label" : {
        "type": "string"
      }, 
      "_id" : {
        "type": "string"
      }
    }
  },
...
   "Category" : {
    "required": [
     
    ],
    "properties": {
    "value": {
      "type": "string"
    }, 
    "label" : {
      "type": "string"
    }
  }
},
...
  }
}